Added simulated component
