+++ acs/sensor_tag/test/CDB/MACI/Components/Components.xml
@@ -6,6 +6,9 @@
       <_ Name="SensorTag"       Code="SensorTagImpl"	
                              Type="IDL:example/Sensors/sensortag:1.0"
                              Container="stContainer" ImplLang="cpp" />
+      <_ Name="SensorTagSim"       Code="SensorTagSim"	
+                             Type="IDL:example/Sensors/sensortag:1.0"
+                             Container="cppSimContainer" ImplLang="cpp" />
       <_ Name="MOUNT1"       Code="acsexmplMountImpl"	
                              Type="IDL:alma/MOUNT_ACS/Mount:1.0"
                              Container="bilboContainer" ImplLang="cpp" />
