+++ acs/sensor_tag/test/CDB/MACI/Containers/cppSimContainer/cppSimContainer.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="ISO-8859-1"?>
+<Container xmlns="urn:schemas-cosylab-com:Container:1.0" 
+           xmlns:cdb="urn:schemas-cosylab-com:CDB:1.0" 
+           xmlns:baci="urn:schemas-cosylab-com:BACI:1.0" 
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
+           xmlns:log="urn:schemas-cosylab-com:LoggingConfig:1.0" 
+           Timeout="20.0" 
+           UseIFR="1" 
+           ManagerRetry="10" 
+           ImplLang="cpp">
+  <Autoload>
+    <cdb:_ string="baci" />
+  </Autoload>
+  <LoggingConfig 
+             centralizedLogger="Log" 
+             minLogLevel="2" 
+             dispatchPacketSize="0" 
+             immediateDispatchLevel="99">
+  </LoggingConfig>
+</Container>
+
+
+
+
