+++ acs/sensor_tag/test/CDB/alma/SensorTagSim/SensorTagSim.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="ISO-8859-1"?>
+<SensorTag xmlns="urn:schemas-cosylab-com:SensorTag:1.0"
+       xmlns:baci="urn:schemas-cosylab-com:BACI:1.0"
+	   xmlns:cdb="urn:schemas-cosylab-com:CDB:1.0"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <temperature />
+  <light />
+  <humidity />
+</SensorTag>
+
